requirements.txt
streamlit
torch
torchvision
Pillow
requests
beautifulsoup4
opencv-python
